Move some types
